requirements.txt
Django==4.1.1
djangorestframework==3.13.1
django-phonenumber-field==7.0.0
phonenumbers==8.12.55
flake8==5.0.4